+++ frontend/wailsjs/go/models.ts
@@ -1,3 +1,24 @@
+export namespace db {
+	
+	export class RecentEntry {
+	    path: string;
+	    filename: string;
+	    openedAt: string;
+	
+	    static createFrom(source: any = {}) {
+	        return new RecentEntry(source);
+	    }
+	
+	    constructor(source: any = {}) {
+	        if ('string' === typeof source) source = JSON.parse(source);
+	        this.path = source["path"];
+	        this.filename = source["filename"];
+	        this.openedAt = source["openedAt"];
+	    }
+	}
+
+}
+
 export namespace main {
 	
 	export class PlaybackInfo {
@@ -34,22 +55,11 @@ export namespace main {
 	        this.clickToTogglePlayback = source["clickToTogglePlayback"];
 	    }
 	}
-	export class RecentEntry {
-	    path: string;
-	    filename: string;
-	    openedAt: string;
+
+}
+
+export namespace thumbs {
 	
-	    static createFrom(source: any = {}) {
-	        return new RecentEntry(source);
-	    }
-	
-	    constructor(source: any = {}) {
-	        if ('string' === typeof source) source = JSON.parse(source);
-	        this.path = source["path"];
-	        this.filename = source["filename"];
-	        this.openedAt = source["openedAt"];
-	    }
-	}
 	export class SeekThumbnailData {
 	    vtt: string;
 	    spritesheetBase64: string;